requirements.txt
fastapi
uvicorn
pydantic
motor
openai
transformers
torch
sentence-transformers
python-dotenv
lama-cpp-python